Renamed Grid Company to Actor
