+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/DataAccess/MeteringPointContext.cs
@@ -13,32 +13,36 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MeteringPoints.Domain.Actors;
 using Energinet.DataHub.MeteringPoints.Domain.EnergySuppliers;
 using Energinet.DataHub.MeteringPoints.Domain.GridAreas;
-using Energinet.DataHub.MeteringPoints.Domain.GridCompanies;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Exchange;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Production;
+using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.Actors;
 using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.EnergySuppliers;
 using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.GridAreas;
-using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.GridCompanies;
 using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.MessageHub;
 using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Infrastructure.InternalCommands;
 using Energinet.DataHub.MeteringPoints.Infrastructure.LocalMessageHub;
 using Energinet.DataHub.MeteringPoints.Infrastructure.Messaging.Idempotency;
 using Energinet.DataHub.MeteringPoints.Infrastructure.Outbox;
+using Energinet.DataHub.MeteringPoints.Infrastructure.Serialization;
 using Microsoft.EntityFrameworkCore;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess
 {
     public class MeteringPointContext : DbContext
     {
-        #nullable disable
-        public MeteringPointContext(DbContextOptions<MeteringPointContext> options)
+        private readonly IJsonSerializer _jsonSerializer;
+
+#nullable disable
+        public MeteringPointContext(DbContextOptions<MeteringPointContext> options, IJsonSerializer jsonSerializer)
             : base(options)
         {
+            _jsonSerializer = jsonSerializer;
         }
 
         public MeteringPointContext()
@@ -65,7 +69,7 @@ namespace Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess
 
         public DbSet<EnergySupplier> EnergySuppliers { get; private set; }
 
-        public DbSet<GridCompany> GridCompanies { get; private set; }
+        public DbSet<Actor> Actors { get; private set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
@@ -82,7 +86,8 @@ namespace Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess
             modelBuilder.ApplyConfiguration(new GridAreaEntityConfiguration());
             modelBuilder.ApplyConfiguration(new MessageHubMessageEntityConfiguration());
             modelBuilder.ApplyConfiguration(new EnergySupplierEntityConfiguration());
-            modelBuilder.ApplyConfiguration(new GridCompanyEntityConfiguration());
+            // TODO: Can this be handled with some proper DI instead?
+            modelBuilder.ApplyConfiguration(new ActorEntityConfiguration(_jsonSerializer));
         }
     }
 }
